requirements.txt
alpaca-py
backtrader[plotting]
logzero
matplotlib==3.2.2
pandas
tqdm
yfinance
